9/5 21.30 -21.50 = 20min finito fix fuoco e fiamme. va tutto meglio di prima! topino rompiscatole.... bruuuuuuu

+++ src/main/java/org/galvanica/math/MetodiArrotondamenti.java
@@ -4,7 +4,7 @@ public class MetodiArrotondamenti {
     public static ScattiMath alimentazioneScattiMath(int scattiAttuali,
                                                      int scattiAlimentazione,
                                                      boolean arrotondaValori,
-                                                     Double primoValoreVolumetrico) {
+                                                     Integer primoValoreVolumetrico) {
         //risposta:  private long restoScatti;
         //           private double moltiplicatoreAlimentazione;
         ScattiMath risposta = new ScattiMath();
